pyproject.toml
[tool.poetry]
name = "extractor"
version = "0.1.0"
description = "Scrape, process, and load reports from Texas AG & other sources."
authors = ["elaine ayo <projects@elaineayo.com>"]

############### PROJECT CLIs
[tool.poetry.scripts]
scrape = 'extractor.cli:main'

############### ADDITIONAL DEPENDENCY GROUPS
# optional dependency groups besides base
[tool.poetry.group.dev]
optional = true

[tool.poetry.group.analytics]
optional = true

############### DEPENDENCY/BUILD INFORMATION
[tool.poetry.dependencies]
python = "^3.9"
click = "*"
python-dotenv = "*"
sqlalchemy = "*"
psycopg2-binary = "*"
mashumaro = "*"
requests = "*"
pendulum = "*"
types-requests = "*"
bs4 = "*"
alembic = "*"
pydantic = "*"
pandas = "*"
s3fs = "*"
sqlmodel = "*"


[tool.poetry.group.dev.dependencies]
coverage = "^5.3"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
black = {version = "22.12.0", allow-prereleases = true}
pre-commit = "^2.16.0"
mypy = "^0.930"
pylint = "2.5.3"

[tool.poetry.group.analytics.dependencies]
jupyter = "*"
notebook = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

############### OTHER TOOL CONFIGURATION
# configuration for other tools
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    ]

[[tool.mypy.overrides]]
module = [
    "yaml",
    "pytz",
    "dateutil"
]
ignore_missing_imports = true

[tool.pycln]
all = true

[tool.black]
target_versions = ['py311']
